+++ test/integration/tests/loadexternal.sh
@@ -33,10 +33,10 @@
 
 source helpers.sh
 
-alg_primary_obj=0x000B
-alg_primary_key=0x0001
-alg_create_obj=0x000B
-alg_create_key=0x0008
+alg_primary_obj=sha256
+alg_primary_key=rsa
+alg_create_obj=sha256
+alg_create_key=hmac
 
 file_primary_key_ctx=context.p_"$alg_primary_obj"_"$alg_primary_key"
 file_loadexternal_key_pub=opu_"$alg_create_obj"_"$alg_create_key"