+++ test/integration/tests/tcti/abrmd/extended-sessions.sh
@@ -36,7 +36,6 @@ source helpers.sh
 alg_primary_obj=sha256
 alg_primary_key=ecc
 alg_create_obj=sha256
-alg_create_key=keyedhash
 alg_pcr_policy=sha1
 
 pcr_ids="0,1,2,3"
@@ -45,10 +44,10 @@ file_pcr_value=pcr.bin
 file_input_data=secret.data
 file_policy=policy.data
 file_primary_key_ctx=context.p_"$alg_primary_obj"_"$alg_primary_key"
-file_unseal_key_pub=opu_"$alg_create_obj"_"$alg_create_key"
-file_unseal_key_priv=opr_"$alg_create_obj"_"$alg_create_key"
-file_unseal_key_ctx=ctx_load_out_"$alg_primary_obj"_"$alg_primary_key"-"$alg_create_obj"_"$alg_create_key"
-file_unseal_key_name=name.load_"$alg_primary_obj"_"$alg_primary_key"-"$alg_create_obj"_"$alg_create_key"
+file_unseal_key_pub=opu_"$alg_create_obj"
+file_unseal_key_priv=opr_"$alg_create_obj"
+file_unseal_key_ctx=ctx_load_out_"$alg_primary_obj"_"$alg_primary_key"-"$alg_create_obj"
+file_unseal_key_name=name.load_"$alg_primary_obj"_"$alg_primary_key"-"$alg_create_obj"
 file_unseal_output_data=usl_"$file_unseal_key_ctx"
 file_session_file="session.dat"
 
@@ -104,8 +103,8 @@ tpm2_policypcr -Q -S $file_session_file -L ${alg_pcr_policy}:${pcr_ids} -F $file
 
 tpm2_flushcontext -S $file_session_file
 
-tpm2_create -Q -g $alg_create_obj -G $alg_create_key -u $file_unseal_key_pub -r $file_unseal_key_priv -I- -C $file_primary_key_ctx -L $file_policy \
-  -A 'sign|fixedtpm|fixedparent|sensitivedataorigin' <<< $secret
+tpm2_create -Q -g $alg_create_obj -u $file_unseal_key_pub -r $file_unseal_key_priv -I- -C $file_primary_key_ctx -L $file_policy \
+  -A 'fixedtpm|fixedparent' <<< $secret
 
 tpm2_load -Q -C $file_primary_key_ctx -u $file_unseal_key_pub -r $file_unseal_key_priv -n $file_unseal_key_name -o $file_unseal_key_ctx